api/pyproject.toml
[tool.poetry]
name = "api_konsi"
version = "0.1.0"
description = ""
authors = ["ViniciusEspindola <vinnyaoe@gmail.com>"]
readme = "README.md"
packages = [{include = "api_konsi"}]

[tool.poetry.dependencies]
python = "^3.11.4"
uvicorn = "^0.18.3"
fastapi = "^0.83.0"
python-decouple = "^3.8"
requests = "^2.31.0"
pika = "^1.3.2"
redis = "^4.6.0"
elasticsearch = "^8.9.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
